Storage business logic refactored
