+++ sdk-java/src/main/java/ly/count/sdk/java/internal/ModuleCrash.java
@@ -42,7 +42,7 @@ public class ModuleCrash extends ModuleBase {
                 Thread.setDefaultUncaughtExceptionHandler(previousHandler);
             }
             if (clear) {
-                ctx.getSDK().storablePurge(ctx, CrashImplCore.getStoragePrefix());
+                ctx.getSDK().clyStorage.storablePurge(ctx, CrashImplCore.getStoragePrefix());
             }
         } catch (Throwable t) {
             L.e("[ModuleCrash] Exception while stopping crash reporting" + t);
