+++ sdk-java/src/main/java/ly/count/sdk/java/internal/ModuleRatingCore.java
@@ -53,7 +53,7 @@ public class ModuleRatingCore extends ModuleBase {
     }
 
     public void PurgeRatingInfo(){
-        ctx.getSDK().storablePurge(ctx, storableStoragePrefix);
+        ctx.getSDK().clyStorage.storablePurge(ctx, storableStoragePrefix);
     }
 
     /**
